backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==1.4
psycopg2-binary
pandas
httpx
python-dotenv
alembic
